Crate private_event_sourcing

Structs

AwaitingDependencies

EncryptedMessage

EventHistory

EventHistorySummary

PrivateEventEntry

SignedContent

SignedEvent

Enums

EntryTypes

LinkTypes

PrivateEventSourcingRemoteSignal

Signal

UnitEntryTypes

Traits

EventType

PrivateEvent

Functions

__num_entry_types

__num_link_types

action_hash

attempt_commit_awaiting_deps_entries

call_local_zome

call_send_events

commit_my_pending_encrypted_messages

create_encrypted_message

create_private_event

entry_defs

genesis_self_check

Validation you perform during the genesis process. Nobody else on the network performs it, only you. There is no access to network calls in this callback

get_all_agents_for

get_linked_devices_for

get_linked_devices_with_proof_for

import_events

init

linked_devices_zome_name

post_commit

private_event_entry_to_signed_event

query_all_my_agents

query_event_histories

query_my_linked_devices

query_my_linked_devices_with_proof

query_private_event

query_private_event_entries

query_private_event_entries_by_type

query_private_event_entry

query_private_events

query_private_events_by_type

receive_private_events

recv_private_events_remote_signal

scheduled_tasks

send_events

send_private_events_to_new_recipients

validate

This is the unified validation callback for all entries and link types in this integrity zome Below is a match template for all of the variants of DHT Ops and entry and link types

validate_agent_joining

Validation the network performs when you try to join, you can’t perform this validation yourself as you are not a member yet. There is access to network calls in this function

validate_create_awaiting_dependencies

validate_create_encrypted_message

validate_create_event_history

validate_create_event_history_summary

validate_create_link_agent_encrypted_message

validate_create_private_event

validate_delete_awaiting_dependencies

validate_delete_encrypted_message

validate_delete_event_history

validate_delete_event_history_summary

validate_delete_link_agent_encrypted_message

validate_delete_private_event

validate_private_event_entry

validate_update_awaiting_dependencies

validate_update_encrypted_message

validate_update_event_history

validate_update_event_history_summary

validate_update_private_event

Attribute Macros

private_event

Derive Macros

IntoStaticStr

Implements From<MyEnum> for &'static str on an enum.