private_event_sourcing
Trait PrivateEvent

pub trait PrivateEvent:
    EventType
    + Clone
    + TryFrom<SerializedBytes>
    + TryInto<SerializedBytes> {
    // Required methods
    fn validate(
        &self,
        event_hash: EntryHash,
        author: AgentPubKey,
        timestamp: Timestamp,
    ) -> ExternResult<ValidateCallbackResult>;
    fn recipients(
        &self,
        event_hash: EntryHash,
        author: AgentPubKey,
        timestamp: Timestamp,
    ) -> ExternResult<BTreeSet<AgentPubKey>>;

    // Provided method
    fn post_commit(
        &self,
        entry_hash: EntryHash,
        author: AgentPubKey,
        timestamp: Timestamp,
    ) -> ExternResult<()> { ... }
}

Required Methods

fn validate( &self, event_hash: EntryHash, author: AgentPubKey, timestamp: Timestamp, ) -> ExternResult<ValidateCallbackResult>

Whether the given entry is to be accepted in to our source chain

fn recipients( &self, event_hash: EntryHash, author: AgentPubKey, timestamp: Timestamp, ) -> ExternResult<BTreeSet<AgentPubKey>>

The agents other than the linked devices for the author that are suposed to receive this entry

Provided Methods

fn post_commit( &self, entry_hash: EntryHash, author: AgentPubKey, timestamp: Timestamp, ) -> ExternResult<()>

Code to run after an event has been committed

Object Safety

This trait is not object safe.