Crate private_event_sourcing

Structs

Acknowledgement

AcknowledgementContent

EventHistory

EventSentToRecipients

EventSentToRecipientsContent

Message

PrivateEventContent

PrivateEventEntry

ReceiveMessageInput

SignedContent

SignedEntry

Enums

AwaitingDependencies

EntryTypes

PrivateEventSourcingRemoteSignal

Signal

UnitEntryTypes

Traits

EventType

PrivateEvent

Functions

__num_entry_types

action_hash

attempt_commit_awaiting_deps_entries

call_local_zome

call_send_events

create_acknowledgements_for

create_pending_acknowledgements

create_private_event

entry_defs

export_event_history

genesis_self_check

Validation you perform during the genesis process. Nobody else on the network performs it, only you. There is no access to network calls in this callback

get_all_agents_for

get_linked_devices_for

get_linked_devices_with_proof_for

import_event_history

init

linked_devices_zome_name

post_commit

private_event_entry_to_signed_event

query_acknowledgement_entries

query_acknowledgements_by_agents

query_all_my_agents

query_event_histories

query_my_acknowledgement_for

query_my_linked_devices

query_my_linked_devices_with_proof

query_private_event

query_private_event_entries

query_private_event_entries_by_type

query_private_event_entry

query_private_events

query_private_events_by_type

receive_acknowledgements

receive_message

receive_private_events

recv_private_events_remote_signal

resend_events_if_necessary

scheduled_tasks

send_acknowledgement_for_event_to_recipient

send_async_message

send_new_events

validate

This is the unified validation callback for all entries and link types in this integrity zome Below is a match template for all of the variants of DHT Ops and entry and link types

validate_agent_joining

Validation the network performs when you try to join, you can’t perform this validation yourself as you are not a member yet. There is access to network calls in this function

validate_create_acknowledgement

validate_create_awaiting_dependencies

validate_create_event_history

validate_create_events_sent_to_recipients

validate_create_private_event

validate_delete_acknowledgement

validate_delete_awaiting_dependencies

validate_delete_event_history

validate_delete_events_sent_to_recipients

validate_delete_private_event

validate_private_event_entry

validate_update_acknowledgement

validate_update_awaiting_dependencies

validate_update_event_history

validate_update_events_sent_to_recipients

validate_update_private_event

Type Aliases

SignedEvent

Attribute Macros

private_event

Derive Macros

IntoStaticStr

Implements From<MyEnum> for &'static str on an enum.