Trait private_event_sourcing::PrivateEvent

pub trait PrivateEvent: TryFrom<SerializedBytes> + TryInto<SerializedBytes> + Clone {
    // Required methods
    fn validate(
        &self,
        author: AgentPubKey,
        timestamp: Timestamp,
    ) -> ExternResult<ValidateCallbackResult>;
    fn recipients(
        &self,
        author: AgentPubKey,
        timestamp: Timestamp,
    ) -> ExternResult<Vec<AgentPubKey>>;

    // Provided method
    fn post_commit(
        &self,
        author: AgentPubKey,
        timestamp: Timestamp,
    ) -> ExternResult<()> { ... }
}

Required Methods

fn validate( &self, author: AgentPubKey, timestamp: Timestamp, ) -> ExternResult<ValidateCallbackResult>

Whether the given entry is to be accepted in to our source chain

fn recipients( &self, author: AgentPubKey, timestamp: Timestamp, ) -> ExternResult<Vec<AgentPubKey>>

The agents other than the linked devices for the author that are suposed to receive this entry

Provided Methods

fn post_commit( &self, author: AgentPubKey, timestamp: Timestamp, ) -> ExternResult<()>

Code to run after an event has been committed

Object Safety

This trait is not object safe.